Fixed a cgame/game mismatch which was introduced in revision 728 and added the warsow 0.61 changes in cg_hud (i guess joki is ok with this). Also renamed the output dir for the modules to racesow

+++ source/gameshared/q_comref.h
@@ -430,8 +430,7 @@ typedef struct
 	qbyte weaponState;
 
 	int inventory[MAX_ITEMS];
-	// racesow : use int for scores cause we use it for time
-	int stats[PS_MAX_STATS];	// fast status bar updates
+	short stats[PS_MAX_STATS];	// fast status bar updates
 	qbyte plrkeys;				// infos on the pressed keys of chased player (self if not chasing)
 } player_state_t;
